+++ src/Catesta.Settings.ps1
@@ -1,0 +1,1 @@
+[version]$script:requiredPSVersion = '5.1.0'