+++ src/Catesta/Resources/Module/src/PSModule.Settings.ps1
@@ -1,2 +1,2 @@
 # specify the minumum required major PowerShell version that the build script should validate
-[int]$script:requiredPSVersion = '5'
+[version]$script:requiredPSVersion = '5.1.0'